requirements.txt
mcp[cli]>=1.0.0
PyGithub>=2.3.0
python-dotenv>=1.0.0
requests>=2.31.0
PyNaCl>=1.5.0
